adding create endpoint to controller

+++ src/main/java/com/blackpaw/vendingmachine/model/VendingMachine.java
@@ -26,5 +26,5 @@ public abstract class VendingMachine {
     @ElementCollection
     @MapKeyColumn(name="coin_type")
     @Column(name= "count")
-    private  Map<Coin, Double> denominations = new LinkedHashMap<>();
+    private  Map<Coin, Integer> denominations = new LinkedHashMap<>();
 }
